Move CachedGpio to tca9555 namespace
